server/requirements.txt
# --- FastAPI & Server ---
fastapi==0.115.2
uvicorn==0.32.0
python-multipart==0.0.9

# --- Database (MongoDB) ---
motor==3.6.0
pymongo==4.9.0
python-dotenv==1.0.1

# --- NLP & ML (Hugging Face + PyTorch) ---
torch==2.5.1
transformers==4.46.1

# --- PDF Processing ---
pdfplumber==0.11.4

# --- Utilities ---
pydantic==2.9.2
langchain==0.3.7
langchain-ollama==0.2.0
langchain-community==0.3.6

# --- Optional (safe versions for compatibility) ---
typing-extensions>=4.12.2
